### 2022-03-21 #4255 @fabian-s
:wave:
@elimillera @sestelo @Palash63
would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
