### 2022-03-22 #4255 @fabian-s
@sestelo @Palash63
would either of you be willing to review this submission for JOSS?
