+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   Advanced-example: Advanced-example.html
   Basic-examples: Basic-examples.html
   Overview: Overview.html
-last_built: 2022-03-16T14:27Z
+last_built: 2022-03-24T21:22Z
 urls:
   reference: https://inceptdk.github.io/adaptr/reference
   article: https://inceptdk.github.io/adaptr/articles
