+++ .github/workflows/downloads-history-plot.yaml
@@ -2,7 +2,8 @@
 
 on:
   schedule:
-  - cron: "0 12 * * *" # run every day at noon UTC
+  - cron: "0 22 * * 1,5" # run Mondays and Fridays at 22 o'clock
+  workflow_dispatch:
 
 
 name: plot-download-history
