### 2022-03-31 #4284 @editorialbot
[REVIEW]: adaptr: an R package for simulating and comparing adaptive clinical trials
**Submitting author:** @agranholm (<a href="http://orcid.org/0000-0001-5799-7655">Anders Granholm</a>)
**Repository:** https://github.com/INCEPTdk/adaptr
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.0.0
