### 2022-03-31 #4284 @fabian-s
👋🏼 @agranholm @elimillera @stmcg

@elimillera @stmcg  please begin by creating checklists for your reviews by entering

```@editorialbot generate my checklist```

as a comment in this thread.

These checklists contain the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines in the top comment in this thread.

Ping me (@fabian-s) if you have any questions/concerns.
