+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   Advanced-example: Advanced-example.html
   Basic-examples: Basic-examples.html
   Overview: Overview.html
-last_built: 2022-03-28T21:26Z
+last_built: 2022-04-05T12:23Z
 urls:
   reference: https://inceptdk.github.io/adaptr/reference
   article: https://inceptdk.github.io/adaptr/articles
