### 2022-04-06 #4284 @stmcg
This is an excellent package. One comment on the checklist item: "State of the field: Do the authors describe how this software compares to other commonly-used packages?"

The authors mention that existing software for simulating adaptive clinical trial designs is either closed source, has limited features, or is based on graphical user interfaces. It would be helpful if the authors state the key additional features of the ``adaptr`` package compared to existing open source software.
