### 2022-04-08 #4284 @fabian-s
> @fabian-s what is the process from here, should we wait for the second review to complete before revising the manuscript file? Thanks!

you can, but you can also go ahead and address this now.
since we coordinate the review process in this thread and have an open line of communication, we don't need to rely on "review rounds" and can tackle issues as they come up.
