### 2022-04-14 #4284 @agranholm
We have now updated the manuscript to respond to the comment by @stmcg . We have expanded the third paragraph of the discussion (first part of the third paragraph), to add a (brief) comparison of  `adaptr` with other open-source alternatives, with reference to a recent systematic review summarising the features of the available software solutions.
