### 2022-04-26 #4284 @fabian-s
[agranholm](https://github.com/agranholm)

At this point could you:

- [ ] Check the archival deposit (e.g., in Zenodo) has the **correct metadata**. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
