### 2022-04-26 #4284 @agranholm
@elimillera Thank you very much!

@fabian-s Excellent!

We have previously made a tagged release on GitHub of the first (and newest version), v.1.0.0: https://github.com/INCEPTdk/adaptr/releases/tag/v1.0.0

This version of the software - v.1.0.0 - has now been archived at Zenodo and metadata should be correct and match the submission. The DOI is: 10.5281/zenodo.6490087, full link: https://doi.org/10.5281/zenodo.6490087

Thank you very much for moving this forward so quickly.
