### 2022-04-26 #4284 @fabian-s
note for JOSS EiCs: this paper exceeds the recommended size for JOSS papers somewhat (~2000 words), but IMO this is adequate here: there's lots of code examples in the text and the paper does not contain any "fluff" that would be easy to remove without losing important content.
