### 2022-04-26 #4284 @danielskatz
@fabian-s - also note that there can only be one @editorialbot command per comment, and it has to be the first thing in the comment
