### 2022-04-26 #4284 @fabian-s
@danielskatz thanks for your help.

I'm sorry to say I missed one tiny thing during proof-reading earlier -- it's

"Planning adaptive trials and comparing adaptive trials designs is complex, however, and require statistical"
instead of "Planning adaptive trials and comparing adaptive trials designs is complex, however, and require**s** statistical", see https://github.com/INCEPTdk/adaptr/pull/3.

Not sure if we can't proceed anyway?
