### 2022-04-26 #4284 @agranholm
Thanks, I’ve merged @fabian-s a pull request, so the typo is corrected now, thanks for catching this. Not sure if I should regenerate the PDF now at this stage using the usual command or if it should be done by one of you?
