### 2022-04-26 #4284 @epiben
Just chipping in; hope that's fine. The sentence has a grammatical error, but it's not that one. It should read, "Planning adaptive trials and comparing adaptive trials designs **are** complex, however, and require statistical". Planning + comparing => third person plural. This, "require" is correct.
