### 2022-04-26 #4284 @danielskatz
@epiben - thanks for your suggestion - Either way would be fine (alt: Planning adaptive trials and comparing adaptive trials designs is complex, however, and requires statistical) but the singular/plural part should be consistent.
@agranholm - can you also change this, and I will still proofread this whole thing later today?
