### 2022-04-27 #4284 @danielskatz
@openjournals/dev @arfon - note that the title in JOSS metadata includes some quote marks that seem to also be in the suggested citation, while the PDF does not have the quote marks.  Any thoughts on how this should be handled?
