### 2022-04-27 #4284 @agranholm
Amazing, thanks!

@danielskatz We stylized the package name as code using backticks throughout the paper, this seems to be what causes the problem in the title in the metadata (and the font in the title appears unchanged despite the formatting) - maybe the easiest solution is just if we remove the backticks around the package name in the title and keep them in the body of the text (where the formatting works)?
