### 2022-04-27 cfcb9fd
Update README with new CITATION info
