+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   Advanced-example: Advanced-example.html
   Basic-examples: Basic-examples.html
   Overview: Overview.html
-last_built: 2022-04-27T08:09Z
+last_built: 2022-04-27T08:12Z
 urls:
   reference: https://inceptdk.github.io/adaptr/reference
   article: https://inceptdk.github.io/adaptr/articles
