### 2022-10-11 #4255 @Palash63
Hi Fabian,

I can review this submission. Thanks!

Best,
Palash Sharma; Ph.D.

--------------------------------------------------------------

*Sincerely Yours,*

*Palash Sharma*

*Graduate Research Assistant*

*Department of BioStatistics*

*University of Kansas Medical Center*

*Cell No : 775-203-7920*

On Mon, Mar 21, 2022 at 8:20 AM Fabian Scheipl ***@***.***>
wrote:

> 👋
> @elimillera <https://github.com/elimillera> @sestelo
> <https://github.com/sestelo> @Palash63 <https://github.com/Palash63>
> would you be willing to review this submission for JOSS? We carry out our
> checklist-driven reviews here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
